snapshot: fix all tests except cli
