+++ src/backy/tests/test_report.py
@@ -1,22 +1,25 @@
+import pytest
+
 from backy.report import ChunkMismatchReport
-from backy.repository import Repository
 from backy.tests import Ellipsis
 
 
-def test_quarantine(tmp_path, log, clock):
-    repo = Repository()
-    repo.add_report(ChunkMismatchReport(b"source", b"target", 3))
+@pytest.mark.skip
+def test_quarantine(tmp_path, repository, log, clock):
+    repository.add_report(ChunkMismatchReport(b"source", b"target", 3))
     with open(
-        (tmp_path / "quarantine" / repo.report_ids[0]).with_suffix(".report")
+        (tmp_path / "quarantine" / repository.report_ids[0]).with_suffix(
+            ".report"
+        )
     ) as report:
         assert (
             Ellipsis(
                 f"""\
-uuid: {repo.report_ids[0]}
+uuid: {repository.report_ids[0]}
+timestamp: 2015-09-01 07:06:47+00:00
 source_hash: 36cd38f49b9afa08222c0dc9ebfe35eb
 target_hash: 42aefbae01d2dfd981f7da7d823d689e
 offset: 3
-timestamp: 2015-09-01 07:06:47+00:00
 traceback: |-
 ...
     File ".../src/backy/rbd/tests/test_quarantine.py", line ..., in test_quarantine
